RTL: Add backpressure state, rare corner case.
Addressing corner case where next length and potencially some data gets
stored in flop ( `flop_data_q` ). In this case we need to create an
extra cycle gap to read this data out of the flop and into the next
itch message length field ( and the next flopped data ).
In order to create this gap we need to apply backpressure to the input
stream.

This commits cover the re-routing of the length field, re-routing of the
flopped data is in latter commit.